requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.21
pandas==2.1.1
numpy<2.0.0
openpyxl==3.1.2
python-multipart==0.0.6
jinja2==3.1.2
httpx==0.25.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
